[ADD] x real client resolvers
